refactor: improve trace type selector component
